refactor: magi-ui
