+++ src/app/components/magi/MagiRec/ClippedRecMel.tsx
@@ -1,28 +1,35 @@
 import type React from "react";
-import useResponsiveSize from "../hooks/useResponsiveSize"; // Import the hook
+import useResponsiveSize from "../../../hooks/useResponsiveSize"; // Import the hook
 
 // Define default dimensions
-const DEFAULT_WIDTH = 250;
-const DEFAULT_HEIGHT = 300;
+const DEFAULT_WIDTH = 310;
+const DEFAULT_HEIGHT = 160;
 
 interface ClippedRectangleProps {
   // width and height are now handled internally
   color: string;
+  // Keep borderColor and borderWidth if they are meant to be customizable,
+  // otherwise remove them and define defaults inside the component.
+  // Assuming they are not needed based on current usage.
+  // borderColor?: string;
+  // borderWidth?: number;
   children?: React.ReactNode;
 }
 
-const ClippedRecBal: React.FC<ClippedRectangleProps> = ({
+// Renaming component to match filename convention
+const ClippedRecMel: React.FC<ClippedRectangleProps> = ({
   color,
   children,
 }) => {
   // Get responsive dimensions from the hook
   const { width, height } = useResponsiveSize(DEFAULT_WIDTH, DEFAULT_HEIGHT);
 
-  const borderColor = "#FF6600";
-  const borderWidth = 4;
-  const clipPath =
-    "polygon(0% 0%, 100% 0%, 100% 80%, 75% 100%, 25% 100%, 0% 80%)";
+  // The clip path for both the outer and inner elements
+  const clipPath = "polygon(20% 0%, 100% 0%, 100% 100%, 0% 100%, 0% 35%)";
+  const borderColor = "#FF6600"; // Default border color
+  const borderWidth = 4; // Default border width
 
+  // Outer element style (this will be the "border")
   const outerStyle: React.CSSProperties = {
     width: `${width}px`, // Use responsive width
     height: `${height}px`, // Use responsive height
@@ -33,6 +40,7 @@ const ClippedRecBal: React.FC<ClippedRectangleProps> = ({
     justifyContent: "center",
   };
 
+  // Inner element style (this will be the content area)
   const innerStyle: React.CSSProperties = {
     width: `${width - borderWidth * 2}px`, // Use responsive width
     height: `${height - borderWidth * 2}px`, // Use responsive height
@@ -50,4 +58,4 @@ const ClippedRecBal: React.FC<ClippedRectangleProps> = ({
   );
 };
 
-export default ClippedRecBal;
+export default ClippedRecMel; // Export with the correct name
